chore: use getContents instead of src to avoid lodash.template
